<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Splits character by '$' and returns the single characters}
\usage{
extract(input)
}
\arguments{
\item{input}{a vector or a list of characters}
}
\value{
a vector of characters
}
\description{
This is an internal function.
}
\details{
Not required
}
\keyword{internal}

||||||| merged common ancestors
=======
\name{extract}
\alias{extract}
\title{Splits character by '$' and returns the single characters}
\usage{
extract(input)
}
\arguments{
  \item{input}{a vector or a list of characters}
}
\value{
a vector of characters
}
\description{
This is an internal function.
}
\details{


Not required
}
\keyword{internal}

>>>>>>> 5870b4aeaa26728dcafe58c3538c59b94540ed10
